backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0
celery==5.3.6
redis==5.0.1
httpx==0.26.0
pytest==8.0.0
playwright==1.41.1
# locust==2.23.1
jinja2==3.1.3
python-multipart==0.0.9
email-validator==2.1.0.post1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
